Just a note...
I am aware of two errors in this workbook you will encounter in the interpreter:

* The first error (ValueError) occurs on line 27 in [12]. The reason this occurs is because you are asking to add Shape twice. It is there already. The reason I am leaving it as-is is because if you remove the request to add it that precedes it in the code, it will run. That request that precedes it was added later as part of a test, but I like that it's there, so if you want to not have the error come up, simply comment out the earlier request (lines 11-30 in [9], the calculation block) - and if the data source already has 'Shape' as a column, just remove it and it should work.

The second error is an HTML formatting error that occurs in the Output portion. That whole area can be commented out if you wish. You have already run the correctly-formed command to display it in HTML, so this is totally redundant. The reason for the error is that 'results' is not defined, which is because we already did this step in pandas (the instruction to do it in that Output module is from the last part of the assignment.)

While I have certainly done my all to isolate the causes of the problems, naturally, there is no guarantee my explanations are fully correct. I will continue to investigate whether my assertions are right; in the meantime, if you notice something wrong, please let me know.